app/requirements.txt
pydantic
fastapi
uvicorn[standard]
python-jose
SQLAlchemy
alembic
passlib[bcrypt] 
python-multipart
email-validator 
httpx
requests
# redis                  
# python-dateutilredis                     
# python-dateutil